+++ bitcoin-regtest/bitcoin-regtest-example-application/src/main/java/org/tbk/bitcoin/regtest/example/BitcoinRegtestExampleApplicationConfig.java
@@ -10,7 +10,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.tbk.bitcoin.zeromq.client.MessagePublishService;
 import org.tbk.electrum.ElectrumClient;
-import org.tbk.electrum.command.LoadWalletParams;
+import org.tbk.electrum.rpc.command.LoadWalletParams;
 import reactor.core.Disposable;
 import reactor.core.publisher.Flux;