+++ incubator/tbk-electrum-daemon-client/tbk-electrum-daemon-client-core/src/integTest/java/org/tbk/electrum/ElectrumDaemonClientContainerTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import org.tbk.electrum.command.*;
+import org.tbk.electrum.rpc.command.*;
 import org.tbk.electrum.model.*;
 import org.tbk.spring.testcontainer.electrumd.ElectrumDaemonContainer;
 import org.tbk.spring.testcontainer.electrumx.ElectrumxContainer;