+++ incubator/tbk-electrum-daemon-client/tbk-electrum-daemon-client-core/src/main/java/org/tbk/electrum/ElectrumClient.java
@@ -2,7 +2,7 @@ package org.tbk.electrum;
 
 import lombok.Builder;
 import lombok.Value;
-import org.tbk.electrum.command.*;
+import org.tbk.electrum.rpc.command.*;
 import org.tbk.electrum.model.*;
 
 import javax.annotation.Nullable;