+++ incubator/tbk-electrum-daemon-client/tbk-electrum-daemon-client-core/src/main/java/org/tbk/electrum/model/SimpleTxHashAndBlockHeight.java
@@ -3,7 +3,7 @@ package org.tbk.electrum.model;
 import lombok.Builder;
 import lombok.NonNull;
 import lombok.Value;
-import org.tbk.electrum.command.AddressHistoryResponse;
+import org.tbk.electrum.rpc.command.AddressHistoryResponse;
 
 @Value
 @Builder